requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.12.1+cu113
torchvision==0.13.1+cu113
sklearn==0.0
wandb==0.12.16
Pillow
matplotlib
torchlm
segmentation-models-pytorch
pandas
git+https://github.com/ildoonet/pytorch-randaugment
prettytable
tqdm
scikit-learn
webcolors
opencv-python
numba
